fix: made trigger words easier to define
