+++ src/app.js
@@ -15,6 +15,8 @@ const client = new Client({
     ],
 });
 
+const triggers = require("./consts/triggers");
+
 console.log("registering the following commands", Object.keys(commands));
 
 client.on("interactionCreate", (interaction) => {
@@ -55,29 +57,12 @@ client.on("messageCreate", (msg) => {
     const tokens = new Set(sanitized.split(" "));
 
     //mega funny ways to make the bot insult everyone on the server based on stupid shit they say:
-    const triggers = {
-        "crazy": "crazy? i was crazy once. they locked me in a room. a rubber room. a rubber room filled with rats and the rats made me crazy.",
-        "toxic": "why you being toxic? personally i would never.",
-        "hi": "no, shut up, go away.",
-        "hello": "no, shut up, go away.",
-        "hey": "no, shut up, go away.",
-        "morning": "quiet, go back to sleep.",
-        "night": "sleep is for the weak pussy.",
-        "gm": "quiet, go back to sleep. U still seem to be too tired to write u lazy fuck.",
-        "gn": "no need to wake up until you start writing properly you lazy fuck.",
-        "imagine": "stop imagining pussy, this is reality.",
-        "bye": "bye! no need to come back!",
-        "food": "no food for you, you fat fuck!",
-        "eat": "no food for you, you fat fuck!",
-        "surströmming": "ew wtf stop opening your legs. you nasty hoe."
-    }
 
-    Object.entries(triggers).forEach(item => {
+    Object.entries(triggers).forEach((item) => {
         if (tokens.has(item[0])) {
-            msg.channel.send(item[1])
+            msg.channel.send(item[1]);
         }
-    })
-    
+    });
 });
 
 (async () => {
