+++ src/consts/triggers.js
@@ -1,0 +1,18 @@
+const triggers = {
+    "crazy": "crazy? i was crazy once. they locked me in a room. a rubber room. a rubber room filled with rats and the rats made me crazy.",
+    "toxic": "why you being toxic? personally i would never.",
+    "hi": "no, shut up, go away.",
+    "hello": "no, shut up, go away.",
+    "hey": "no, shut up, go away.",
+    "morning": "quiet, go back to sleep.",
+    "night": "sleep is for the weak pussy.",
+    "gm": "quiet, go back to sleep. U still seem to be too tired to write u lazy fuck.",
+    "gn": "no need to wake up until you start writing properly you lazy fuck.",
+    "imagine": "stop imagining pussy, this is reality.",
+    "bye": "bye! no need to come back!",
+    "food": "no food for you, you fat fuck!",
+    "eat": "no food for you, you fat fuck!",
+    "surströmming": "ew wtf stop opening your legs. you nasty hoe."
+};
+
+module.exports = triggers;
